// repos/konraddb/ui_private/src/store/rates/rates.types.ts
export type RateData = {
  id: string;
  base: string;
  quote: string;
  pair: string;
  fullName: string;
  bidPriceValue: number;
  bidPriceUnformatted: string;
  askPriceValue: number;
  askPriceUnformatted: string;
  marketPriceValue: number;
  bidPrice: string;
  askPrice: string;
  marketPrice: string;
  timestamp: string;
};